+++ tests/test_blueprints.py
@@ -971,7 +971,7 @@ def test_nesting_subdomains(app, client) -> None:
     assert response.status_code == 200
 
 
-def test_child_overrides_parent_subdomain(app, client) -> None:
+def test_child_and_parent_subdomain(app, client) -> None:
     child_subdomain = "api"
     parent_subdomain = "parent"
     parent = flask.Blueprint("parent", __name__)
@@ -988,7 +988,9 @@ def test_child_overrides_parent_subdomain(app, client) -> None:
 
     domain_name = "domain.tld"
     app.config["SERVER_NAME"] = domain_name
-    response = client.get("/", base_url=f"http://{child_subdomain}.{domain_name}")
+    response = client.get(
+        "/", base_url=f"http://{child_subdomain}.{parent_subdomain}.{domain_name}"
+    )
 
     assert response.status_code == 200