+++ CHANGES.rst
@@ -3,6 +3,8 @@ Version 2.3.0
 
 Unreleased
 
+-   Ensure subdomains are applied with nested blueprints. :issue:`4834`
+
 
 Version 2.2.3
 -------------